Accept fractional volume and pace market data requests
The first live provider check showed fractional-share volume in vendor
aggregates and HTTP 429 without Retry-After on the free tier. Volume is now
an exact Decimal end to end (migration 0005 makes snapshot volume double
precision), and requests are paced with a fallback wait on bare 429s.

+++ technical/models.py
@@ -65,7 +65,7 @@ class TechnicalSnapshot:
     vwap: float
     rsi: float
     atr: float
-    volume: int
+    volume: float        # Bar volume (may be fractional; see market_data.models).
     average_volume: float
     relative_volume: float
     trend: str
